`rptest`: add cloud topics leveling scale test
Rename `cloud_topics_compaction_scale_test.py` to
`cloud_topics_maintenance_scale_test.py` and extract a shared
`CloudTopicsMaintenanceScaleTest` base. The existing compaction soak
test's behavior is unchanged.

To this file, add `CloudTopicsLevelingScaleTest`, a sustained soak that
fans writes thinly across many partitions (plain cloud topics) with a
raised leveling concurrency cap, asserting leveling stays live under
load, converges once writes stop, reclaims a meaningful number of
extents, and leaves the cluster healthy.

Also, point the gcp suite at the renamed file.
