+++ tests/rptest/test_suite_gcp_temporary.yml
@@ -4,7 +4,7 @@ gcp:
     - tests
     - transactions
     - tuner_integration_tests/net_tuner_test.py::GcpNetTunerTest
-    - scale_tests/cloud_topics_compaction_scale_test.py
+    - scale_tests/cloud_topics_maintenance_scale_test.py
     - scale_tests/many_partitions_test.py::ManyPartitionsTest.test_many_partitions_cloud_topics
     - scale_tests/many_partitions_test.py::ManyPartitionsTest.test_many_partitions_cloud_topics_tiered_storage
     - scale_tests/large_messages_test.py::LargeMessagesTest.test_cloud_topics_large_messages_throughput
